Add support for "new" authentication method
Venmo killed or broke the OAuth authentication method (it now 404s). There is however a "new" authentication method which can be used by passing in the user's username and password 😱.

This change switches the authentication method to use the "new" one. It is documented here https://github.com/mmohades/VenmoApiDocumentation